create an ecnrypted_data_2 column in table users to make the migration from v5 to v6 safer

+++ src/api1/routes/deleteSharing.ts
@@ -12,6 +12,13 @@ export const deleteSharing = async (req: any, res: any): Promise<void> => {
     const basicAuth = await checkBasicAuth(req, { checkIsManagerForItemId: itemId });
     if (!basicAuth.granted) return res.status(401).end();
 
+
+
+    const hasDataV2Res = await db.query("SELECT length(encrypted_data_2) AS data2_length FROM users WHERE id=$1", [basicAuth.userId]);
+    if(hasDataV2Res.rows[0].data2_length > 0) {
+      return res.status(403).json({error: 'deprecated_app'});
+    }
+
     // NB shared_account_users will be deleted by cascade
     await db.query('DELETE FROM shared_accounts WHERE id=$1 AND group_id=$2', [
       itemId,
