+++ src/api1/routes/sendStats.ts
@@ -12,6 +12,13 @@ export const sendStats = async (req: any, res: any): Promise<void> => {
     const basicAuth = await checkBasicAuth(req);
     if (!basicAuth.granted) return res.status(401).end();
 
+
+
+    const hasDataV2Res = await db.query("SELECT length(encrypted_data_2) AS data2_length FROM users WHERE id=$1", [basicAuth.userId]);
+    if(hasDataV2Res.rows[0].data2_length > 0) {
+      return res.status(403).json({error: 'deprecated_app'});
+    }
+
     // remove previous stats this same day
     await db.query(
       "DELETE FROM data_stats WHERE user_id=$1 AND date_trunc('day', date)=date_trunc('day', now()) AND group_id=$2",
